requirements.txt
certifi>=2018.11.29
chardet>=3.0.4
Click>=7.0
idna>=2.8
Jinja2>=2.10
ldap3>=2.5.2
livereload>=2.6.0
lxml>=4.3.1
Markdown>=3.0.1
MarkupSafe>=1.1.0
mkdocs>=1.0.4
pyasn1>=0.4.5
pydoc-markdown>=2.0.5
python-dotenv>=0.10.1
PyYAML>=3.13
requests>=2.21.0
six>=1.12.0
tornado>=5.1.1
urllib3>=1.24.1
responses>=0.10.5